begin removing try/catch from news crud

+++ app/Http/Controllers/API/AdminController.php
@@ -11,7 +11,7 @@ use App\Models\Post;
 use Auth;
 use JWTAuth;
 use Hash;
-
+use App\Http\Requests\CRUDRequest;
 class AdminController extends Controller
 {
 
@@ -33,25 +33,22 @@ class AdminController extends Controller
      * @return Response
      */
     public function editPost(Request $request) {
-        try {
-            $post = Post::findOrFail($request->id);
-            $validator = Validator::make($request->all(), [
-                'title' => 'required|string',
-                'type'   => 'required|integer|between:0,2',
-                'content' => 'required'
-            ]);
-            if ($validator->fails()) {
-                return ['message' => 'validation', 'errors' => $validator->errors()->all()];
-            }
-            $post->title = $request->title;
-            $post->type = $request->type;
-            $post->content = $request->content;
-            $post->save();
-            return response()->json(['message' => 'success'], 200);
-        }
-        catch(ModelNotFoundException $e) {
-            return response()->json(['message' => 'invalid_id'], 401);
-        }
+    	$request['id'] = $request->id;
+		$validator = Validator::make($request->all(), [
+			'id'	=> 'required|exists:posts,id',
+		    'title' => 'required|string',
+		    'type'   => 'required|integer|between:0,2',
+		    'content' => 'required'
+		]);
+		if ($validator->fails()) {
+		    return ['message' => 'validation', 'errors' => $validator->errors()->all()];
+		}
+		$post = Post::find($request->id);
+		$post->title = $request->title;
+		$post->type = $request->type;
+		$post->content = $request->content;
+		$post->save();
+		return response()->json(['message' => 'success'], 200);
     }
 
     /**
@@ -61,14 +58,16 @@ class AdminController extends Controller
      * @return Response
      */
     public function deletePost(Request $request) {
-        try {
-            $post = Post::findOrFail($request->id);
-            $post->delete();
-            return response()->json(['message' => 'success'], 200);
-        }
-        catch(ModelNotFoundException $e) {
-            return response()->json(['message' => 'invalid_id'], 401);
-        }
+    	$request['id'] = $request->id;
+		$validator = Validator::make($request->all(), [
+			'id'	=> 'required|exists:posts,id',
+		]);
+		if ($validator->fails()) {
+		    return ['message' => 'validation', 'errors' => $validator->errors()->all()];
+		}
+		$post = Post::find($request->id);
+        $post->delete();
+        return response()->json(['message' => 'success'], 200);
     }
 
     /**
